## Change Summary

Removed a `.compute()` call from `test_open_dataset_kerchunk`. It's been bugging me for a couple of days that the tests seemed to be running excessively slowly.

The kerchunk file that was being computed is approximate 2GB in size & we shouldn't need to have a `.compute()` call to assert that it is a dataset (AFAIK).

This reduces test suite execution time from:
-  ~220s to ~120s on my machine
-  ~130s to 105s on CI


## Checklist

- [x] Unit tests for the changes exist N/A
- [x] Tests pass on CI 
- [x] Documentation reflects the changes where applicable N/A 

<!--
Please add any other relevant info below:
-->